requirements/base.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
alabaster==0.7.16
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
autoflake==2.3.1
Automat==25.4.16
autopep8==2.3.2
babel==2.17.0
billiard==4.2.1
black==25.1.0
boto3==1.28.64
botocore==1.31.85
Brotli==1.1.0
celery==5.3.4
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.3
cssselect2==0.8.0
daphne==4.0.0
distlib==0.3.9
dj-database-url==2.1.0
Django==4.2.7
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.0
django-debug-toolbar==4.3.0
django-filter==24.3
django-redis==5.4.0
django-storages==1.14.2
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework-gis==1.1
djangorestframework-simplejwt==5.3.0
docutils==0.18.1
drf-yasg==1.21.7
ffmpeg-python==0.2.0
filelock==3.18.0
flake8==7.2.0
fonttools==4.57.0
frozenlist==1.6.0
future==1.0.0
geographiclib==2.0
geopy==2.4.0
html5lib==1.1
hyperlink==21.0.0
identify==2.6.10
idna==3.10
imagesize==1.4.1
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.5.3
MarkupSafe==3.0.2
mccabe==0.7.0
msgpack==1.1.0
multidict==6.4.3
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
phonenumbers==8.13.22
pillow==11.2.1
platformdirs==4.3.8
pre_commit==4.2.0
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydyf==0.11.0
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyphen==0.17.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.2
redis==5.0.1
requests==2.32.3
s3transfer==0.7.0
service-identity==24.2.0
six==1.17.0
snowballstemmer==2.2.0
Sphinx==6.2.1
sphinx-rtd-theme==1.2.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
tinycss2==1.4.0
twilio==8.9.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
virtualenv==20.31.1
wcwidth==0.2.13
weasyprint==60.1
webencodings==0.5.1
whitenoise==6.5.0
yamllint==1.37.1
yarl==1.20.0
zope.interface==7.2
zopfli==0.2.3.post1
